+++ docs/whats_new.rst
@@ -73,6 +73,7 @@ Enhancements
 - Only load data if needed during preprocessing (e.g., allow timecrop without loading) (:gh:`164` by `Robin Tibor Schirrmeister`_)
 - Adding option to sort filtered channels by frequency band for the filterbank in :func:`braindecode.datautil.filterbank` (:gh:`185` by `Lukas Gemein`_)
 - Adding the USleep model :class:`braindecode.models.USleep` (:gh:`282` by `Theo Gnassounou`_ and `Omar Chehab`_)
+- Adding SleepStagerEldele2021 and SleepStagerBlanco2020 models :class:`braindecode.models.SleepStagerEldele2021` and :class:`braindecode.models.SleepStagerBlanco2020` (:gh:`341` by `Divyesh Narayanan`_)
 
 Bugs
 ~~~~
@@ -105,6 +106,7 @@ Authors
 .. _Vytautas Jankauskas: https://github.com/vytjan
 .. _Theo Gnassounou: https://github.com/Tgnassou
 .. _Omar Chehab: https://github.com/l-omar-chehab
+.. _Divyesh Narayanan: https://github.com/Div12345
 .. _Alex Gramfort: http://alexandre.gramfort.net
 .. _Cedric Rommel: https://cedricrommel.github.io
 .. _Simon Brandt: https://github.com/sbbrandt